requirements.txt
httpx==0.23.3
pymongo==4.3.3
python-dotenv==1.0.0
PyYAML==6.0
